Main (skeleton) Assembler File
This file is an incapsulated source code of virus for better debuging.
Files with .inc are include files with adittional code that declares
functions and data.
There are some troubles, the code works, but some parts are
misunderstood by me, so now it's kind of magic for me. (need to be
solved)