Added working client part.
Now it can try to connect to server when started and wait for server's
answer. When the connection is established, client waits for commands
from server. We can ask it to send us something, to finish itself and to
close connection and start listening. The last function isn't working
yet.